+++ internal/provider/provider.go
@@ -193,7 +193,7 @@ func (p *EmporixProvider) Resources(ctx context.Context) []func() resource.Resou
 		NewDeliveryTimeResource,
 		NewShippingMethodResource,
 		NewTaxResource,
-		NewPriceModelResource,
+		NewPriceModelsResource,
 	}
 }